NeMo Data Designer/backend2/requirements.txt
fastapi
uvicorn[standard]
pandas
python-dotenv
nemo-microservices[data-designer]